Published new version 1.2.2
git-svn-id: https://plugins.svn.wordpress.org/assetsminify/trunk@833449 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ AssetsMinifyInit.php
@@ -106,7 +106,12 @@ class AssetsMinifyInit {
 	 */
 	public function gc() {
 		update_option( 'am_last_gc', time() );
-		foreach ( glob("{$this->assetsPath}*.*") as $filepath )
+
+		$files = glob("{$this->assetsPath}*.*");
+		if ( $files === false )
+			return false;
+
+		foreach ( $files  as $filepath )
 			if ( filemtime($filepath) <= time() - 864000 ) //If the file is older than 10 days then is removed
 				unlink($filepath);
 	}
@@ -430,7 +435,11 @@ class AssetsMinifyInit {
 
 		//Prints <script> inclusion in the page
 		$this->dumpScriptData( 'footer' );
-		$this->dumpJs( "foot-{$mtime}.js" );
+
+		$async = false;
+		if( get_option('am_async_flag', 1) )
+			$async = true;
+		$this->dumpJs( "foot-{$mtime}.js", $async );
 	}
 
 	/**
